+++ frontend/src/routes.tsx
@@ -1,7 +1,7 @@
 import type { RouteObject } from "react-router"
 import AllRoleListing from "./pages/applicant/AllRoleListing"
 import ViewRoleListing from "./pages/applicant/main"
-import Signup from "./pages/hr/signup"
+import Signup from "./pages/hr/CreateRoleListing"
 
 const routes: RouteObject[] = [
   {
